+++ crates/swc_ecma_minifier/src/compress/pure/strings.rs
@@ -4,10 +4,166 @@ use swc_atoms::{js_word, JsWord};
 use swc_common::{util::take::Take, DUMMY_SP};
 use swc_ecma_ast::*;
 use swc_ecma_utils::{ExprExt, Type, Value};
+use Value::Known;
 
 use super::Pure;
 
 impl Pure<'_> {
+    /// This only handles `'foo' + ('bar' + baz) because others are handled by
+    /// expression simplifier.
+    pub(super) fn eval_str_addition(&mut self, e: &mut Expr) {
+        let (span, l_l, r_l, r_r) = match e {
+            Expr::Bin(
+                e @ BinExpr {
+                    op: op!(bin, "+"), ..
+                },
+            ) => match &mut *e.right {
+                Expr::Bin(
+                    r @ BinExpr {
+                        op: op!(bin, "+"), ..
+                    },
+                ) => (e.span, &mut *e.left, &mut *r.left, &mut r.right),
+                _ => return,
+            },
+            _ => return,
+        };
+
+        let lls = l_l.as_string();
+        let rls = r_l.as_string();
+
+        if let (Known(lls), Known(rls)) = (lls, rls) {
+            self.changed = true;
+            report_change!("evaluate: 'foo' + ('bar' + baz) => 'foobar' + baz");
+
+            let s = lls.into_owned() + &*rls;
+            *e = Expr::Bin(BinExpr {
+                span,
+                op: op!(bin, "+"),
+                left: s.into(),
+                right: r_r.take(),
+            });
+        }
+    }
+
+    pub(super) fn eval_tpl_as_str(&mut self, e: &mut Expr) {
+        if !self.options.evaluate || !self.options.unsafe_passes {
+            return;
+        }
+
+        let tpl = match e {
+            Expr::Tpl(e) => e,
+            _ => return,
+        };
+
+        if tpl.quasis.len() == 2
+            && (tpl.quasis[0].cooked.is_some() || !tpl.quasis[0].raw.contains('\\'))
+            && tpl.quasis[1].raw.is_empty()
+        {
+            self.changed = true;
+            report_change!("evaluating a template to a string");
+            *e = Expr::Bin(BinExpr {
+                span: tpl.span,
+                op: op!(bin, "+"),
+                left: tpl.quasis[0]
+                    .cooked
+                    .clone()
+                    .unwrap_or_else(|| tpl.quasis[0].raw.clone())
+                    .into(),
+                right: tpl.exprs[0].take(),
+            });
+        }
+    }
+
+    pub(super) fn eval_nested_tpl(&mut self, e: &mut Expr) {
+        let tpl = match e {
+            Expr::Tpl(e) => e,
+            _ => return,
+        };
+
+        if !tpl.exprs.iter().any(|e| match &**e {
+            Expr::Tpl(t) => t
+                .quasis
+                .iter()
+                .all(|q| q.cooked.is_some() || !q.raw.contains('\\')),
+            _ => false,
+        }) {
+            return;
+        }
+
+        self.changed = true;
+        report_change!("evaluating nested template literals");
+
+        let mut new_tpl = Tpl {
+            span: tpl.span,
+            quasis: Default::default(),
+            exprs: Default::default(),
+        };
+        let mut cur_str_value = String::new();
+
+        for idx in 0..(tpl.quasis.len() + tpl.exprs.len()) {
+            if idx % 2 == 0 {
+                let q = tpl.quasis[idx / 2].take();
+
+                cur_str_value.push_str(q.cooked.as_deref().unwrap_or(&*q.raw));
+            } else {
+                let mut e = tpl.exprs[idx / 2].take();
+                self.eval_nested_tpl(&mut *e);
+
+                match *e {
+                    Expr::Tpl(mut e) => {
+                        // We loop again
+                        //
+                        // I think we can merge this code...
+                        for idx in 0..(e.quasis.len() + e.exprs.len()) {
+                            if idx % 2 == 0 {
+                                let q = e.quasis[idx / 2].take();
+
+                                cur_str_value.push_str(q.cooked.as_deref().unwrap_or(&*q.raw));
+                            } else {
+                                let s = JsWord::from(&*cur_str_value);
+                                cur_str_value.clear();
+
+                                new_tpl.quasis.push(TplElement {
+                                    span: DUMMY_SP,
+                                    tail: false,
+                                    cooked: Some(s.clone()),
+                                    raw: s,
+                                });
+
+                                let e = e.exprs[idx / 2].take();
+
+                                new_tpl.exprs.push(e);
+                            }
+                        }
+                    }
+                    _ => {
+                        let s = JsWord::from(&*cur_str_value);
+                        cur_str_value.clear();
+
+                        new_tpl.quasis.push(TplElement {
+                            span: DUMMY_SP,
+                            tail: false,
+                            cooked: Some(s.clone()),
+                            raw: s,
+                        });
+
+                        new_tpl.exprs.push(e);
+                    }
+                }
+            }
+        }
+
+        let s = JsWord::from(&*cur_str_value);
+        new_tpl.quasis.push(TplElement {
+            span: DUMMY_SP,
+            tail: false,
+            cooked: Some(s.clone()),
+            raw: s,
+        });
+
+        *e = Expr::Tpl(new_tpl);
+    }
+
     /// Converts template literals to string if `exprs` of [Tpl] is empty.
     pub(super) fn convert_tpl_to_str(&mut self, e: &mut Expr) {
         match e {
